refactor: restructure component
